requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
defusedxml==0.7.1
Django==4.0.6
djangorestframework==3.13.1
httpie==3.2.1
idna==3.3
kombu==5.2.4
multidict==6.0.2
prompt-toolkit==3.0.30
Pygments==2.12.0
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
